countercurvature/requirements.txt
numpy>=2.3.0
pandas>=2.2.0
matplotlib>=3.9.0
PyYAML>=6.0.0
biopython>=1.84

